Move ResultExt to std-utils
