+++ utils/std-utils/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "ya-std-utils"
+version = "0.1.0"
+authors = ["Golem Factory <contact@golem.network>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+log = "0.4"
