+++ utils/std-utils/src/lib.rs
@@ -1,0 +1,3 @@
+mod result;
+
+pub use result::ResultExt;
